feat: renamed KonnectorItem to KonnectorTile 📝
