+++ src/styles/konnectorTile.styl
